requirements.txt
scipy
iexfinance
numpy
matplotlib
seaborn
pandas
tabulate
texttable
requests
python-dotenv
python-twitter
tweepy
python-environ
beautifulsoup4
python-telegram-bot
jsonfield
redis
colored
pytrends
python-dateutil
progressbar2
django
django-tailwind
django-compressor 
django-compressor-toolkit
django-sass
praw
cryptography
dateparser